Use FLATMEM to allow booting with memory cgroups on
commit 6422531ae224d73b55e8fa095d1c7405651eabac
Author: Jani Monoses <jani@ubuntu.com>
Date: Mon Apr 29 16:31:27 2013 +0300

Use FLATMEM to allow booting with memory cgroups on

The memory cgroups setup when sparsemem is enabled is not
working well with the layout of the mako hardware memory
blocks and causes a page fault when booting. This works around
the bug which may be either in the cgroups or mako hw setup
code.

Change-Id:I3bf34bc5e2cff25e613595bc8bff06c81b711198
Signed-off-by: Jani Monoses<jani@ubuntu.com>

bug refer:
[1] http://permalink.gmane.org/gmane.linux.kernel.cgroups/6590
[2] http://permalink.gmane.org/gmane.linux.kernel.cgroups/7003

Signed-off-by: Qiu Feng <qiufeng@xiaomi.com>